+++ android/src/main/kotlin/rekab/app/background_locator/BackgroundLocatorPlugin.kt
@@ -10,9 +10,6 @@ import android.os.Build
 import android.os.Handler
 import android.util.Log
 import androidx.core.content.ContextCompat
-import com.google.android.gms.location.LocationRequest
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
@@ -22,68 +19,23 @@ import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry
-import io.flutter.view.FlutterMain
-import rekab.app.background_locator.Keys.Companion.ARG_CALLBACK
-import rekab.app.background_locator.Keys.Companion.ARG_CALLBACK_DISPATCHER
-import rekab.app.background_locator.Keys.Companion.ARG_DISPOSE_CALLBACK
-import rekab.app.background_locator.Keys.Companion.ARG_INIT_CALLBACK
-import rekab.app.background_locator.Keys.Companion.ARG_INIT_DATA_CALLBACK
-import rekab.app.background_locator.Keys.Companion.ARG_NOTIFICATION_CALLBACK
-import rekab.app.background_locator.Keys.Companion.ARG_SETTINGS
-import rekab.app.background_locator.Keys.Companion.BACKGROUND_CHANNEL_ID
-import rekab.app.background_locator.Keys.Companion.BCM_NOTIFICATION_CLICK
-import rekab.app.background_locator.Keys.Companion.CALLBACK_DISPATCHER_HANDLE_KEY
-import rekab.app.background_locator.Keys.Companion.CALLBACK_HANDLE_KEY
-import rekab.app.background_locator.Keys.Companion.CHANNEL_ID
-import rekab.app.background_locator.Keys.Companion.DISPOSE_CALLBACK_HANDLE_KEY
-import rekab.app.background_locator.Keys.Companion.INIT_CALLBACK_HANDLE_KEY
-import rekab.app.background_locator.Keys.Companion.INIT_DATA_CALLBACK_KEY
-import rekab.app.background_locator.Keys.Companion.METHOD_PLUGIN_INITIALIZE_SERVICE
-import rekab.app.background_locator.Keys.Companion.METHOD_PLUGIN_IS_REGISTER_LOCATION_UPDATE
-import rekab.app.background_locator.Keys.Companion.METHOD_PLUGIN_IS_SERVICE_RUNNING
-import rekab.app.background_locator.Keys.Companion.METHOD_PLUGIN_REGISTER_LOCATION_UPDATE
-import rekab.app.background_locator.Keys.Companion.METHOD_PLUGIN_UN_REGISTER_LOCATION_UPDATE
-import rekab.app.background_locator.Keys.Companion.METHOD_PLUGIN_UPDATE_NOTIFICATION
-import rekab.app.background_locator.Keys.Companion.NOTIFICATION_ACTION
-import rekab.app.background_locator.Keys.Companion.NOTIFICATION_CALLBACK_HANDLE_KEY
-import rekab.app.background_locator.Keys.Companion.SETTINGS_ACCURACY
-import rekab.app.background_locator.Keys.Companion.SETTINGS_ANDROID_NOTIFICATION_BIG_MSG
-import rekab.app.background_locator.Keys.Companion.SETTINGS_ANDROID_NOTIFICATION_CHANNEL_NAME
-import rekab.app.background_locator.Keys.Companion.SETTINGS_ANDROID_NOTIFICATION_ICON
-import rekab.app.background_locator.Keys.Companion.SETTINGS_ANDROID_NOTIFICATION_ICON_COLOR
-import rekab.app.background_locator.Keys.Companion.SETTINGS_ANDROID_NOTIFICATION_MSG
-import rekab.app.background_locator.Keys.Companion.SETTINGS_ANDROID_NOTIFICATION_TITLE
-import rekab.app.background_locator.Keys.Companion.SETTINGS_ANDROID_WAKE_LOCK_TIME
-import rekab.app.background_locator.Keys.Companion.SETTINGS_DISTANCE_FILTER
-import rekab.app.background_locator.Keys.Companion.SETTINGS_INTERVAL
-import rekab.app.background_locator.Keys.Companion.SHARED_PREFERENCES_KEY
-import rekab.app.background_locator.provider.*
-import java.util.HashMap
-
 
 class BackgroundLocatorPlugin
-    : MethodCallHandler, FlutterPlugin, PluginRegistry.NewIntentListener, ActivityAware, LocationUpdateListener {
-    private lateinit var locatorClient: BLLocationProvider
+    : MethodCallHandler, FlutterPlugin, PluginRegistry.NewIntentListener, ActivityAware {
     private var context: Context? = null
     private var activity: Activity? = null
 
     companion object {
-        @JvmStatic
-        private var isLocationServiceRunning: Boolean = false
-
         @JvmStatic
         private var channel: MethodChannel? = null
 
         @SuppressLint("MissingPermission")
         @JvmStatic
         private fun registerLocator(context: Context,
-                                    client: BLLocationProvider,
                                     args: Map<Any, Any>,
                                     result: Result?) {
-            if (IsolateHolderService.isRunning) {
+            if (PreferencesManager.isServiceRunning(context)) {
                 // The service is running already
-                isLocationServiceRunning = true
-
                 Log.d("BackgroundLocatorPlugin", "Locator service is already running")
                 return
             }
@@ -91,21 +43,13 @@ class BackgroundLocatorPlugin
             Log.d("BackgroundLocatorPlugin",
                     "start locator with ${PreferencesManager.getLocationClient(context)} client")
 
-            isLocationServiceRunning = true
+            val callbackHandle = args[Keys.ARG_CALLBACK] as Long
+            setCallbackHandle(context, Keys.CALLBACK_HANDLE_KEY, callbackHandle)
 
-            val callbackHandle = args[ARG_CALLBACK] as Long
-            setCallbackHandle(context, CALLBACK_HANDLE_KEY, callbackHandle)
+            val notificationCallback = args[Keys.ARG_NOTIFICATION_CALLBACK] as? Long
+            setCallbackHandle(context, Keys.NOTIFICATION_CALLBACK_HANDLE_KEY, notificationCallback)
 
-            val notificationCallback = args[ARG_NOTIFICATION_CALLBACK] as? Long
-            setCallbackHandle(context, NOTIFICATION_CALLBACK_HANDLE_KEY, notificationCallback)
-
-            val initCallback = args[ARG_INIT_CALLBACK] as? Long
-            setCallbackHandle(context, INIT_CALLBACK_HANDLE_KEY, initCallback)
-
-            val disposeCallback = args[ARG_DISPOSE_CALLBACK] as? Long
-            setCallbackHandle(context, DISPOSE_CALLBACK_HANDLE_KEY, disposeCallback)
-
-            val settings = args[ARG_SETTINGS] as Map<*, *>
+            val settings = args[Keys.ARG_SETTINGS] as Map<*, *>
 
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M &&
                     context.checkSelfPermission(Manifest.permission.ACCESS_FINE_LOCATION)
@@ -115,28 +59,34 @@ class BackgroundLocatorPlugin
                 result?.error(msg, null, null)
             }
 
-            val initialDataMap = args[ARG_INIT_DATA_CALLBACK] as? Map<*, *>
-            setDataCallback(context, INIT_DATA_CALLBACK_KEY, initialDataMap)
             startIsolateService(context, settings)
-
-            client.requestLocationUpdates(getLocationRequest(settings))
-
             result?.success(true)
+            PreferencesManager.setServiceRunning(context, true)
         }
 
         @JvmStatic
         private fun startIsolateService(context: Context, settings: Map<*, *>) {
             val intent = Intent(context, IsolateHolderService::class.java)
             intent.action = IsolateHolderService.ACTION_START
-            intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_CHANNEL_NAME, settings[SETTINGS_ANDROID_NOTIFICATION_CHANNEL_NAME] as String)
-            intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_TITLE, settings[SETTINGS_ANDROID_NOTIFICATION_TITLE] as String)
-            intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_MSG, settings[SETTINGS_ANDROID_NOTIFICATION_MSG] as String)
-            intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_BIG_MSG, settings[SETTINGS_ANDROID_NOTIFICATION_BIG_MSG] as String)
-            intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_ICON, settings[SETTINGS_ANDROID_NOTIFICATION_ICON] as String)
-            intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_ICON_COLOR, settings[SETTINGS_ANDROID_NOTIFICATION_ICON_COLOR] as Long)
+            intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_CHANNEL_NAME,
+                    settings[Keys.SETTINGS_ANDROID_NOTIFICATION_CHANNEL_NAME] as String)
+            intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_TITLE,
+                    settings[Keys.SETTINGS_ANDROID_NOTIFICATION_TITLE] as String)
+            intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_MSG,
+                    settings[Keys.SETTINGS_ANDROID_NOTIFICATION_MSG] as String)
+            intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_BIG_MSG,
+                    settings[Keys.SETTINGS_ANDROID_NOTIFICATION_BIG_MSG] as String)
+            intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_ICON,
+                    settings[Keys.SETTINGS_ANDROID_NOTIFICATION_ICON] as String)
+            intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_ICON_COLOR,
+                    settings[Keys.SETTINGS_ANDROID_NOTIFICATION_ICON_COLOR] as Long)
+            intent.putExtra(Keys.SETTINGS_INTERVAL, settings[Keys.SETTINGS_INTERVAL] as Int)
+            intent.putExtra(Keys.SETTINGS_ACCURACY, settings[Keys.SETTINGS_ACCURACY] as Int)
+            intent.putExtra(Keys.SETTINGS_DISTANCE_FILTER, settings[Keys.SETTINGS_DISTANCE_FILTER] as Double)
 
-            if (settings.containsKey(SETTINGS_ANDROID_WAKE_LOCK_TIME)) {
-                intent.putExtra(SETTINGS_ANDROID_WAKE_LOCK_TIME, settings[SETTINGS_ANDROID_WAKE_LOCK_TIME] as Int)
+            if (settings.containsKey(Keys.SETTINGS_ANDROID_WAKE_LOCK_TIME)) {
+                intent.putExtra(Keys.SETTINGS_ANDROID_WAKE_LOCK_TIME,
+                        settings[Keys.SETTINGS_ANDROID_WAKE_LOCK_TIME] as Int)
             }
 
             ContextCompat.startForegroundService(context, intent)
@@ -151,59 +101,25 @@ class BackgroundLocatorPlugin
 
         @JvmStatic
         private fun initializeService(context: Context, args: Map<Any, Any>) {
-            val callbackHandle: Long = args[ARG_CALLBACK_DISPATCHER] as Long
+            val callbackHandle: Long = args[Keys.ARG_CALLBACK_DISPATCHER] as Long
             setCallbackDispatcherHandle(context, callbackHandle)
         }
 
         @JvmStatic
-        private fun getLocationRequest(settings: Map<*, *>): LocationRequestOptions {
-            val interval: Long = (settings[SETTINGS_INTERVAL] as Int * 1000).toLong()
-            val accuracyKey = settings[SETTINGS_ACCURACY] as Int
-            val accuracy = getAccuracy(accuracyKey)
-            val distanceFilter = settings[SETTINGS_DISTANCE_FILTER] as Double
-
-            return LocationRequestOptions(interval, accuracy, distanceFilter.toFloat())
-        }
-
-        @JvmStatic
-        private fun getAccuracy(key: Int): Int {
-            return when (key) {
-                0 -> LocationRequest.PRIORITY_NO_POWER
-                1 -> LocationRequest.PRIORITY_LOW_POWER
-                2 -> LocationRequest.PRIORITY_BALANCED_POWER_ACCURACY
-                3 -> LocationRequest.PRIORITY_BALANCED_POWER_ACCURACY
-                4 -> LocationRequest.PRIORITY_HIGH_ACCURACY
-                else -> LocationRequest.PRIORITY_HIGH_ACCURACY
-            }
-        }
-
-        @JvmStatic
-        private fun removeLocator(context: Context,
-                                  client: BLLocationProvider) {
-            isLocationServiceRunning = false
-            if (!IsolateHolderService.isRunning) {
+        private fun unRegisterPlugin(context: Context) {
+            if (!PreferencesManager.isServiceRunning(context)) {
                 // The service is not running
                 Log.d("BackgroundLocatorPlugin", "Locator service is not running, nothing to stop")
                 return
             }
 
-            client.removeLocationUpdates()
+            PreferencesManager.setServiceRunning(context, false)
             stopIsolateService(context)
         }
 
         @JvmStatic
-        private fun isRegisterLocator(result: Result?) {
-            if (IsolateHolderService.isRunning) {
-                result?.success(true)
-            } else {
-                result?.success(false)
-            }
-            return
-        }
-
-        @JvmStatic
-        private fun isServiceRunning(result: Result?) {
-            if (isLocationServiceRunning) {
+        private fun isServiceRunning(context: Context, result: Result?) {
+            if (PreferencesManager.isServiceRunning(context)) {
                 result?.success(true)
             } else {
                 result?.success(false)
@@ -215,17 +131,17 @@ class BackgroundLocatorPlugin
         private fun updateNotificationText(context: Context, args: Map<Any, Any>) {
             val intent = Intent(context, IsolateHolderService::class.java)
             intent.action = IsolateHolderService.ACTION_UPDATE_NOTIFICATION
-            if (args.containsKey(SETTINGS_ANDROID_NOTIFICATION_TITLE)) {
-                intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_TITLE,
-                        args[SETTINGS_ANDROID_NOTIFICATION_TITLE] as String)
+            if (args.containsKey(Keys.SETTINGS_ANDROID_NOTIFICATION_TITLE)) {
+                intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_TITLE,
+                        args[Keys.SETTINGS_ANDROID_NOTIFICATION_TITLE] as String)
             }
-            if (args.containsKey(SETTINGS_ANDROID_NOTIFICATION_MSG)) {
-                intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_MSG,
-                        args[SETTINGS_ANDROID_NOTIFICATION_MSG] as String)
+            if (args.containsKey(Keys.SETTINGS_ANDROID_NOTIFICATION_MSG)) {
+                intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_MSG,
+                        args[Keys.SETTINGS_ANDROID_NOTIFICATION_MSG] as String)
             }
-            if (args.containsKey(SETTINGS_ANDROID_NOTIFICATION_BIG_MSG)) {
-                intent.putExtra(SETTINGS_ANDROID_NOTIFICATION_BIG_MSG,
-                        args[SETTINGS_ANDROID_NOTIFICATION_BIG_MSG] as String)
+            if (args.containsKey(Keys.SETTINGS_ANDROID_NOTIFICATION_BIG_MSG)) {
+                intent.putExtra(Keys.SETTINGS_ANDROID_NOTIFICATION_BIG_MSG,
+                        args[Keys.SETTINGS_ANDROID_NOTIFICATION_BIG_MSG] as String)
             }
 
             ContextCompat.startForegroundService(context, intent)
@@ -233,57 +149,33 @@ class BackgroundLocatorPlugin
 
         @JvmStatic
         private fun setCallbackDispatcherHandle(context: Context, handle: Long) {
-            context.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
+            context.getSharedPreferences(Keys.SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
                     .edit()
-                    .putLong(CALLBACK_DISPATCHER_HANDLE_KEY, handle)
+                    .putLong(Keys.CALLBACK_DISPATCHER_HANDLE_KEY, handle)
                     .apply()
         }
 
         @JvmStatic
         fun setCallbackHandle(context: Context, key: String, handle: Long?) {
             if (handle == null) {
-                context.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
+                context.getSharedPreferences(Keys.SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
                         .edit()
                         .remove(key)
                         .apply()
                 return
             }
 
-            context.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
+            context.getSharedPreferences(Keys.SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
                     .edit()
                     .putLong(key, handle)
                     .apply()
         }
 
         @JvmStatic
-        fun setDataCallback(context: Context, key: String, data: Map<*, *>?) {
-            if (data == null) {
-                context.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
-                        .edit()
-                        .remove(key)
-                        .apply()
-                return
-            }
-            val dataStr = Gson().toJson(data)
-            context.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
-                    .edit()
-                    .putString(key, dataStr)
-                    .apply()
-        }
-
-        @JvmStatic
         fun getCallbackHandle(context: Context, key: String): Long? {
-            val sharedPreferences = context.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
+            val sharedPreferences = context.getSharedPreferences(Keys.SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
             if (sharedPreferences.contains(key)) return sharedPreferences.getLong(key, 0L)
             return null
-        }
-
-        @JvmStatic
-        fun getDataCallback(context: Context, key: String): Map<*, *> {
-            val initialDataStr = context.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
-                    .getString(key, null)
-            val type = object : TypeToken<Map<*, *>>() {}.type
-            return Gson().fromJson(initialDataStr, type)
         }
 
         @JvmStatic
@@ -292,27 +184,15 @@ class BackgroundLocatorPlugin
 
             val plugin = BackgroundLocatorPlugin()
             plugin.context = context
-            plugin.locatorClient = plugin.getLocationClient(context)
 
             initializeService(context, settings)
-            registerLocator(context,
-                    plugin.locatorClient,
-                    settings, null)
-        }
-
-
-    }
-
-    fun getLocationClient(context: Context): BLLocationProvider {
-        return when (PreferencesManager.getLocationClient(context)) {
-            LocationClient.Google -> GoogleLocationProviderClient(context, this)
-            LocationClient.Android -> AndroidLocationProviderClient(context, this)
+            startIsolateService(context, settings)
         }
     }
 
     override fun onMethodCall(call: MethodCall, result: Result) {
         when (call.method) {
-            METHOD_PLUGIN_INITIALIZE_SERVICE -> {
+            Keys.METHOD_PLUGIN_INITIALIZE_SERVICE -> {
                 val args: Map<Any, Any> = call.arguments()
 
                 // save callback dispatcher to use it when device reboots
@@ -321,26 +201,27 @@ class BackgroundLocatorPlugin
                 initializeService(context!!, args)
                 result.success(true)
             }
-            METHOD_PLUGIN_REGISTER_LOCATION_UPDATE -> {
+            Keys.METHOD_PLUGIN_REGISTER_LOCATION_UPDATE -> {
                 val args: Map<Any, Any> = call.arguments()
 
                 // save setting to use it when device reboots
                 PreferencesManager.saveSettings(context!!, args)
 
-                locatorClient = getLocationClient(context!!)
                 registerLocator(context!!,
-                        locatorClient,
                         args,
                         result)
             }
-            METHOD_PLUGIN_UN_REGISTER_LOCATION_UPDATE -> {
-                removeLocator(context!!,
-                        locatorClient)
+            Keys.METHOD_PLUGIN_UN_REGISTER_LOCATION_UPDATE -> {
+                unRegisterPlugin(context!!)
                 result.success(true)
             }
-            METHOD_PLUGIN_IS_REGISTER_LOCATION_UPDATE -> isRegisterLocator(result)
-            METHOD_PLUGIN_IS_SERVICE_RUNNING -> isServiceRunning(result)
-            METHOD_PLUGIN_UPDATE_NOTIFICATION -> {
+            Keys.METHOD_PLUGIN_IS_REGISTER_LOCATION_UPDATE -> isServiceRunning(context!!, result)
+            Keys.METHOD_PLUGIN_IS_SERVICE_RUNNING -> isServiceRunning(context!!, result)
+            Keys.METHOD_PLUGIN_UPDATE_NOTIFICATION -> {
+                if (!PreferencesManager.isServiceRunning(context!!)) {
+                    return
+                }
+
                 val args: Map<Any, Any> = call.arguments()
                 updateNotificationText(context!!, args)
                 result.success(true)
@@ -359,27 +240,26 @@ class BackgroundLocatorPlugin
     private fun onAttachedToEngine(context: Context, messenger: BinaryMessenger) {
         val plugin = BackgroundLocatorPlugin()
         plugin.context = context
-        plugin.locatorClient = getLocationClient(context)
 
-        channel = MethodChannel(messenger, CHANNEL_ID)
+        channel = MethodChannel(messenger, Keys.CHANNEL_ID)
         channel?.setMethodCallHandler(plugin)
     }
 
     override fun onNewIntent(intent: Intent?): Boolean {
-        if (intent?.action != NOTIFICATION_ACTION) {
+        if (intent?.action != Keys.NOTIFICATION_ACTION) {
             // this is not our notification
             return false
         }
 
-        val notificationCallback = getCallbackHandle(activity!!, NOTIFICATION_CALLBACK_HANDLE_KEY)
-        if (notificationCallback != null && IsolateHolderService.backgroundFlutterView != null) {
-            val backgroundChannel = MethodChannel(IsolateHolderService.backgroundFlutterView,
-                    BACKGROUND_CHANNEL_ID)
+        val notificationCallback = getCallbackHandle(activity!!, Keys.NOTIFICATION_CALLBACK_HANDLE_KEY)
+        if (notificationCallback != null && IsolateHolderService.backgroundEngine != null) {
+            val backgroundChannel =
+                    MethodChannel(IsolateHolderService.backgroundEngine?.dartExecutor?.binaryMessenger, Keys.BACKGROUND_CHANNEL_ID)
             activity?.mainLooper?.let {
                 Handler(it)
                         .post {
-                            backgroundChannel.invokeMethod(BCM_NOTIFICATION_CLICK,
-                                    hashMapOf(ARG_NOTIFICATION_CALLBACK to notificationCallback))
+                            backgroundChannel.invokeMethod(Keys.BCM_NOTIFICATION_CLICK,
+                                    hashMapOf(Keys.ARG_NOTIFICATION_CALLBACK to notificationCallback))
                         }
             }
         }
@@ -401,41 +281,5 @@ class BackgroundLocatorPlugin
     override fun onDetachedFromActivityForConfigChanges() {
     }
 
-    override fun onLocationUpdated(location: HashMap<Any, Any>?) {
-        FlutterMain.ensureInitializationComplete(context!!, null)
 
-        //https://github.com/flutter/plugins/pull/1641
-        //https://github.com/flutter/flutter/issues/36059
-        //https://github.com/flutter/plugins/pull/1641/commits/4358fbba3327f1fa75bc40df503ca5341fdbb77d
-        // new version of flutter can not invoke method from background thread
-        if (location != null) {
-            val callback = getCallbackHandle(context!!, CALLBACK_HANDLE_KEY) as Long
-
-            val result: HashMap<Any, Any> =
-                    hashMapOf(ARG_CALLBACK to callback,
-                            Keys.ARG_LOCATION to location)
-
-            synchronized(IsolateHolderService.serviceStarted) {
-                sendLocationEvent(result)
-            }
-        }
-
-    }
-
-    private fun sendLocationEvent(result: HashMap<Any, Any>) {
-        //https://github.com/flutter/plugins/pull/1641
-        //https://github.com/flutter/flutter/issues/36059
-        //https://github.com/flutter/plugins/pull/1641/commits/4358fbba3327f1fa75bc40df503ca5341fdbb77d
-        // new version of flutter can not invoke method from background thread
-
-        if (IsolateHolderService.backgroundFlutterView != null) {
-            val backgroundChannel = MethodChannel(IsolateHolderService.backgroundFlutterView,
-                    BACKGROUND_CHANNEL_ID)
-            Handler(context!!.mainLooper)
-                    .post {
-                        Log.d("plugin", "sendLocationEvent $result")
-                        backgroundChannel.invokeMethod(Keys.BCM_SEND_LOCATION, result)
-                    }
-        }
-    }
 }