+++ android/src/main/kotlin/rekab/app/background_locator/PreferencesManager.kt
@@ -157,5 +157,18 @@ class PreferencesManager {
             val client = sharedPreferences.getInt(Keys.SETTINGS_ANDROID_LOCATION_CLIENT, 0)
             return LocationClient.fromInt(client) ?: LocationClient.Google
         }
+
+        @JvmStatic
+        fun isServiceRunning(context: Context): Boolean {
+            return context.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE)
+                    .getBoolean(Keys.PREF_SERVICE_IS_RUNNING, false)
+        }
+
+        @JvmStatic
+        fun setServiceRunning(context: Context, running: Boolean) {
+            context.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE)
+                    .edit().putBoolean(Keys.PREF_SERVICE_IS_RUNNING, running)
+                    .apply()
+        }
     }
 }