+++ android/src/main/kotlin/rekab/app/background_locator/provider/BLLocationProvider.kt
@@ -1,9 +1,5 @@
 package rekab.app.background_locator.provider
 
-import android.app.PendingIntent
-import android.content.Intent
-import java.util.HashMap
-
 interface BLLocationProvider {
     var listener: LocationUpdateListener?