+++ ios/Classes/Globals.m
@@ -11,9 +11,6 @@
 
 NSString *const kCallbackDispatcherKey = @"callback_dispatcher_handle_key";
 NSString *const kCallbackKey = @"callback_handle_key";
-NSString *const kInitCallbackKey = @"init_callback_handle_key";
-NSString *const kInitDataCallbackKey = @"init_data_callback_key";
-NSString *const kDisposeCallbackKey = @"dispose_callback_handle_key";
 NSString *const kDistanceFilterKey = @"distance_filter_key";
 
 NSString *const kChannelId = @"app.rekab/locator_plugin";
@@ -48,7 +45,9 @@ NSString *const kSettingsDistanceFilter = @"settings_distanceFilter";
 NSString *const kSettingsShowsBackgroundLocationIndicator = @"settings_ios_showsBackgroundLocationIndicator";
 
 NSString *const kBCMSendLocation = @"BCM_SEND_LOCATION";
-NSString *const kBCMInit = @"BCM_INIT";
-NSString *const kBCMDispose = @"BCM_DISPOSE";
+
+NSString *const kPrefObservingRegion = @"pref_observingRegion";
+NSString *const kPrefServiceRunning = @"pref_serviceRunning";
+
 
 @end