+++ ios/Classes/Preferences/PreferencesManager.h
@@ -17,6 +17,10 @@ NS_ASSUME_NONNULL_BEGIN
 + (void)setCallbackHandle:(int64_t)handle key:(NSString *)key;
 + (void)saveDistanceFilter:(double) distance;
 + (double)getDistanceFilter;
++ (void)setObservingRegion:(BOOL) observing;
++ (BOOL)isObservingRegion;
++ (void)setServiceRunning:(BOOL) running;
++ (BOOL)isServiceRunning;
 
 @end